Organized Imports. Moving private class into PausableThreadPoolExecutor.
