+++ src/main/java/com/marklogic/developer/corb/AbstractTask.java
@@ -22,29 +22,20 @@ import static com.marklogic.developer.corb.Manager.DEFAULT_BATCH_URI_DELIM;
 import static com.marklogic.developer.corb.Manager.URIS_BATCH_REF;
 import static com.marklogic.developer.corb.Options.BATCH_URI_DELIM;
 import static com.marklogic.developer.corb.Options.ERROR_FILE_NAME;
-import static com.marklogic.developer.corb.Options.QUERY_RETRY_LIMIT;
-import static com.marklogic.developer.corb.Options.QUERY_RETRY_INTERVAL;
 import static com.marklogic.developer.corb.Options.QUERY_RETRY_ERROR_CODES;
 import static com.marklogic.developer.corb.Options.QUERY_RETRY_ERROR_MESSAGE;
+import static com.marklogic.developer.corb.Options.QUERY_RETRY_INTERVAL;
+import static com.marklogic.developer.corb.Options.QUERY_RETRY_LIMIT;
 import static com.marklogic.developer.corb.Options.XCC_CONNECTION_RETRY_INTERVAL;
 import static com.marklogic.developer.corb.Options.XCC_CONNECTION_RETRY_LIMIT;
-import com.marklogic.developer.corb.util.StringUtils;
+import static com.marklogic.developer.corb.TransformOptions.FAILED_URI_TOKEN;
 import static com.marklogic.developer.corb.util.StringUtils.commaSeparatedValuesToList;
 import static com.marklogic.developer.corb.util.StringUtils.isEmpty;
 import static com.marklogic.developer.corb.util.StringUtils.isNotEmpty;
 import static com.marklogic.developer.corb.util.StringUtils.trim;
-import com.marklogic.xcc.ContentSource;
-import com.marklogic.xcc.Request;
-import com.marklogic.xcc.RequestOptions;
-import com.marklogic.xcc.ResultSequence;
-import com.marklogic.xcc.Session;
-import com.marklogic.xcc.exceptions.QueryException;
-import com.marklogic.xcc.exceptions.RequestException;
-import com.marklogic.xcc.exceptions.RequestPermissionException;
-import com.marklogic.xcc.exceptions.RetryableQueryException;
-import com.marklogic.xcc.exceptions.ServerConnectionException;
-import com.marklogic.xcc.types.XdmBinary;
-import com.marklogic.xcc.types.XdmItem;
+import static java.util.logging.Level.SEVERE;
+import static java.util.logging.Level.WARNING;
+
 import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
@@ -58,10 +49,21 @@ import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
 import java.util.TimeZone;
-import static java.util.logging.Level.SEVERE;
-import static java.util.logging.Level.WARNING;
 import java.util.logging.Logger;
-import static com.marklogic.developer.corb.TransformOptions.FAILED_URI_TOKEN;
+
+import com.marklogic.developer.corb.util.StringUtils;
+import com.marklogic.xcc.ContentSource;
+import com.marklogic.xcc.Request;
+import com.marklogic.xcc.RequestOptions;
+import com.marklogic.xcc.ResultSequence;
+import com.marklogic.xcc.Session;
+import com.marklogic.xcc.exceptions.QueryException;
+import com.marklogic.xcc.exceptions.RequestException;
+import com.marklogic.xcc.exceptions.RequestPermissionException;
+import com.marklogic.xcc.exceptions.RetryableQueryException;
+import com.marklogic.xcc.exceptions.ServerConnectionException;
+import com.marklogic.xcc.types.XdmBinary;
+import com.marklogic.xcc.types.XdmItem;
 
 /**
  *
