+++ src/main/java/com/marklogic/developer/corb/PausableThreadPoolExecutor.java
@@ -19,8 +19,10 @@
 package com.marklogic.developer.corb;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.TreeSet;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.Executors;
 import java.util.concurrent.RejectedExecutionHandler;
@@ -175,6 +177,79 @@ public class PausableThreadPoolExecutor extends ThreadPoolExecutor {
 	 */
 	public int getNumSucceededUris() {
 		return numSucceededUris;
+	}
+	private class TopUriList{
+		private TreeSet<UriObject>  list = null;
+		private int size=0;
+		
+		public TopUriList(int size) {
+			this.size = size;
+			list = new TreeSet<UriObject>() {
+				private static final long serialVersionUID = 1L;
+				public String toString() {
+					StringBuffer strBuff = new StringBuffer();
+					for (UriObject o : this) {
+						strBuff.append(o.toString());
+					}
+					return strBuff.toString();
+				}
+			};
+		}
+		Map<String,Long> getData(){
+			Map<String,Long> map = new HashMap<String,Long>();
+			for(UriObject obj:this.list){
+				map.put(obj.uri, obj.timeTaken);
+			}
+			return map;
+		}
+		void add(String uri, Long timeTaken){
+			UriObject newObj=new UriObject(uri, timeTaken);
+			if(list.size()<this.size || list.last().compareTo(newObj) <1){
+				synchronized (list) {
+					if(list.size()>=this.size ){
+						for(int i=0; i<=list.size()-this.size; i++){
+							list.remove(list.first());						
+						}
+					}
+					list.add(newObj);
+				}
+			}
+		}
+		private class UriObject implements Comparable<UriObject>{
+			@Override
+			public String toString() {
+				return "UriObject [uri=" + uri + ", timeTaken=" + timeTaken + "]";
+			}
+			String uri;
+			Long timeTaken;
+			public UriObject(String uri, Long timeTaken) {
+				super();
+				this.uri = uri;
+				this.timeTaken = timeTaken;
+			}
+			@Override
+			public boolean equals(Object obj) {
+				if(obj instanceof UriObject){
+					UriObject o = (UriObject)obj;
+					if(this.timeTaken!=null && o.timeTaken!=null) {
+						return this.timeTaken.compareTo(o.timeTaken) == 0;
+					}
+					else{
+						return false;
+					}
+				}
+				else return super.equals(obj);
+			}
+			@Override
+			public int compareTo(UriObject o) {
+				if(this.timeTaken!=null && o.timeTaken!=null) {
+					return this.timeTaken.compareTo(o.timeTaken);
+				}
+				else{
+					return 0;//should never get here
+				}
+			}
+		}
 	}
 }
